The Metropolis Algorithm

The Metropolis Algorithm (Metropolis et al. 1953)

M(startValue = NULL, iterations = 10000, nBI = 0, parmin = NULL,
  parmax = NULL, f = 1, FUN, consoleUpdates = 1000)

Arguments

startValue
vector with the start values for the algorithm. Can be NULL if FUN is of class BayesianSetup. In this case startValues are sampled from the prior.

iterations
iterations to run

nBI
number of burnin

parmin
minimum values for the parameter vector or NULL if FUN is of class BayesianSetup

parmax
maximum values for the parameter vector or NULL if FUN is of class BayesianSetup

f
scaling factor

FUN
function to be sampled from or object of class bayesianSetup

consoleUpdates
interger, determines the frequency with which sampler progress is printed to the console

References

Metropolis, Nicholas, et al. "Equation of state calculations by fast computing machines." The journal of chemical physics 21.6 (1953): 1087-1092.