T-walk MCMC

T-walk MCMC

Twalk(bayesianSetup, settings = list(iterations = 10000, at = 6, aw = 1.5, pn1
  = NULL, Ptrav = 0.4918, Pwalk = 0.4918, Pblow = 0.0082, burnin = 0, thin = 1,
  startValue = NULL, consoleUpdates = 100, message = TRUE))

Arguments

bayesianSetup
Object of class 'bayesianSetup' or 'bayesianOuput'.

settings
list with parameter values.

iterations
Number of model evaluations

at
"traverse" move proposal parameter. Default to 6

aw
"walk" move proposal parameter. Default to 1.5

pn1
Probability determining the number of parameters that are changed

Ptrav
Move probability of "traverse" moves, default to 0.4918

Pwalk
Move probability of "walk" moves, default to 0.4918

Pblow
Move probability of "traverse" moves, default to 0.0082

burnin
number of iterations treated as burn-in. These iterations are not recorded in the chain.

thin
thinning parameter. Determines the interval in which values are recorded.

startValue
Matrix with start values

consoleUpdates
Intervall in which the sampling progress is printed to the console

message
logical determines whether the sampler's progress should be printed

Value

Object of class bayesianOutput.

Details

The probability of "hop" moves is 1 minus the sum of all other probabilities.

References

Christen, J. Andres, and Colin Fox. "A general purpose sampling algorithm for continuous distributions (the t-walk)." Bayesian Analysis 5.2 (2010): 263-281.