Creates a DHARMa object

Creates a DHARMa object

getDharmaResiduals(model, parMatrix, numSamples, observed, error, plot = TRUE)

Arguments

model
function that calculates model predictions for a given parameter vector

parMatrix
a parameter matrix from which the simulations will be generated

numSamples
the number of samples

observed
a vector of observed values

error
function with signature f(mean, par) that generates error expectations from mean model predictions. Par is a vector from the matrix with the parameter samples (full length). f needs to know which of these parameters are parameters of the error function

plot
logical, determining whether the simulated residuals should be plotted